Move implicit-related helpers from typer too the object, too
