+++ effekt/shared/src/main/scala/effekt/Typer.scala
@@ -1450,81 +1450,6 @@ object Typer extends Phase[NameResolved, Typechecked] {
     }
   }
 
-  def instantiateImplicitBlock(b: source.Term, tpe: symbols.BlockType)(using Context): source.Term = {
-    if(!Context.messaging.hasErrors) {
-      (b, tpe) match {
-        case (a, symbols.BlockType.FunctionType(tps, cps, vps, bps, res, effs)) =>
-          a match {
-            case source.BlockLiteral(tparams, vparams, bparams, source.Return(source.Call(fn, targs, vargs, bargs, _), _), _) =>
-              // We need to refresh the whole binding structure, so we don't have duplicate stuff in the tree.
-              // Doing this in a very specialized way here.
-              // It annotates the correct concrete types for *this* invocation.
-              val ftpsyms = tparams.map { x => symbols.TypeParam(Name.local(x.name)) }
-              val ftparams = (tparams zip ftpsyms).map { (x, sym) =>
-                val r = source.IdDef(x.name, source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, sym)
-                r
-              }
-              val ftargs = ftpsyms.map { x =>
-                val r = source.TypeRef(source.IdRef(Nil, x.name.name, source.Span.missing), Many(Nil, source.Span.missing), source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, x)
-                r
-              }
-              val fvpsyms = (vparams zip vps).map { (x, t) => symbols.ValueParam(Name.local(x.id.name), Some(t), false, NoSource) }
-              val fvparams = (vparams zip fvpsyms).map { (x, sym) =>
-                val r: source.ValueParam = source.ValueParam(source.IdDef(x.id.name, source.Span.missing), Some(source.ReifiedType(sym.tpe.get)), false, source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, sym)
-                Context.annotate(Annotations.Symbol, r.id, sym)
-                r
-              }
-              val fvargs = fvpsyms.map { x =>
-                val r = source.Var(source.IdRef(Nil, x.name.name, source.Span.missing), source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, x)
-                Context.annotate(Annotations.Symbol, r.id, x)
-                source.ValueArg(None, r, source.Span.missing)
-              }
-              val fbpsyms = (bparams zip bps).map { (x, t) => symbols.BlockParam(Name.local(x.id.name), Some(t), x.symbol.capture, false, NoSource) }
-              val fbparams = (bparams zip fbpsyms).map { (x, sym) =>
-                val r: source.BlockParam = source.BlockParam(source.IdDef(x.id.name, source.Span.missing), Some(source.ReifiedType(sym.tpe.get)), false, source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, sym)
-                Context.annotate(Annotations.Symbol, r.id, sym)
-                r
-              }
-              val fbargs = fbpsyms.map { x =>
-                val r = source.Var(source.IdRef(Nil, x.name.name, source.Span.missing), source.Span.missing)
-                Context.annotate(Annotations.Symbol, r, x)
-                Context.annotate(Annotations.Symbol, r.id, x)
-                r
-              }
-              val ffn = fn match {
-                case source.IdTarget(id) =>
-                  val r = source.IdTarget(source.IdRef(Nil, id.name, source.Span.missing))
-                  Context.annotate(Annotations.Symbol, r.id,
-                    id.symbol match {
-                      case symbols.CallTarget(syms, impls) =>
-                        symbols.CallTarget(syms, impls) // needs to be refreshed for recursive uses
-                    })
-                  r
-                case _ => Context.panic("Implicit block argument should be an (eta-expanded) name, not an expression")
-              }
-              source.BlockLiteral(ftparams, fvparams, fbparams,
-                source.Return(source.Call(ffn, ftargs, fvargs, fbargs,
-                  source.Span.missing), source.Span.missing), source.Span.missing)
-            case _ => Context.panic("Unexpected implicit value for implicit block parameter")
-          }
-        case (a, symbols.BlockType.InterfaceType(tCons, tArgs)) =>
-          // There is nothing to do here
-          a
-      }
-    } else {
-      Context.abort("Not instantiating implicit block argument since there are errors.")
-    }
-  }
-  def instantiateImplicitValue(v: source.ValueArg, tpe: symbols.ValueType)(using Context): source.ValueArg = {
-    // There is nothing to do here
-    v
-  }
-
   def checkCallTo(
     call: source.CallLike,
     name: String,
@@ -1607,9 +1532,10 @@ object Typer extends Phase[NameResolved, Typechecked] {
       effs = effs ++ eff.toEffects
     }
 
+    // implicit arguments work like normal ones, except that we first have to instantiate them,
+    // and later annotate them to be inserted
     (implicitVps zip implicitVargs) foreach { case (tpe, expr) =>
-      // TODO NOW refresh and instantiate them
-      val inst = instantiateImplicitValue(expr, tpe)
+      val inst = source.GenerateImplicitArgs.instantiateImplicitValue(expr, tpe)
       instImplicitVargs.append(inst)
       val Result(t, eff) = checkExpr(inst.value, Some(tpe))
       effs = effs ++ eff.toEffects
@@ -1635,18 +1561,20 @@ object Typer extends Phase[NameResolved, Typechecked] {
       }
     }
 
+    // implicit arguments work like normal ones, except that we first have to instantiate them,
+    // and...
     (implicitBps zip (implicitBargs zip implicitCaptArgs)) foreach { case (tpe, (expr, capt)) =>
-      // TODO NOW, refresh and instantiate before checking
       flowsInto(capt, callsite)
       // capture of block <: ?C
       flowingInto(capt) {
-        val inst = instantiateImplicitBlock(expr, tpe)
+        val inst = source.GenerateImplicitArgs.instantiateImplicitBlock(expr, tpe)
         instImplicitBargs.append(inst)
         val Result(t, eff) = checkExprAsBlock(inst, Some(tpe))
         effs = effs ++ eff.toEffects
       }
     }
 
+    // ... annotate them to be inserted by [[ExplicitCapabilites]]
     Context.annotateImplicits(call, instImplicitVargs.toList, instImplicitBargs.toList)
 
     // We add return effects last to have more information at this point to
