+++ effekt/shared/src/main/scala/effekt/source/GenerateImplicitArgs.scala
@@ -1,10 +1,13 @@
-package effekt.source
+package effekt
+package source
+
 import scala.collection.mutable
 import effekt.util.messages.{ErrorReporter, EffektMessages}
 import effekt.context.Context
 import effekt.symbols
 import effekt.symbols.scopes.Scope
-import effekt.symbols.{BlockSymbol, BlockType, Callable, ImplicitContext, builtins}
+import effekt.symbols.{BlockSymbol, BlockType, Callable, ImplicitContext, builtins, Name}
+import effekt.context.Annotations
 
 object GenerateImplicitArgs {
 
@@ -75,4 +78,81 @@ object GenerateImplicitArgs {
       }
     }.toMap
   }
+
+
+  def instantiateImplicitBlock(b: source.Term, tpe: symbols.BlockType)(using Context): source.Term = {
+    if(!Context.messaging.hasErrors) {
+      (b, tpe) match {
+        case (a, symbols.BlockType.FunctionType(tps, cps, vps, bps, res, effs)) =>
+          a match {
+            case source.BlockLiteral(tparams, vparams, bparams, source.Return(source.Call(fn, targs, vargs, bargs, _), _), _) =>
+              // We need to refresh the whole binding structure, so we don't have duplicate stuff in the tree.
+              // Doing this in a very specialized way here.
+              // It annotates the correct concrete types for *this* invocation.
+              val ftpsyms = tparams.map { x => symbols.TypeParam(Name.local(x.name)) }
+              val ftparams = (tparams zip ftpsyms).map { (x, sym) =>
+                val r = source.IdDef(x.name, source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, sym)
+                r
+              }
+              val ftargs = ftpsyms.map { x =>
+                val r = source.TypeRef(source.IdRef(Nil, x.name.name, source.Span.missing), Many(Nil, source.Span.missing), source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, x)
+                r
+              }
+              val fvpsyms = (vparams zip vps).map { (x, t) => symbols.ValueParam(Name.local(x.id.name), Some(t), false, NoSource) }
+              val fvparams = (vparams zip fvpsyms).map { (x, sym) =>
+                val r: source.ValueParam = source.ValueParam(source.IdDef(x.id.name, source.Span.missing), Some(source.ReifiedType(sym.tpe.get)), false, source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, sym)
+                Context.annotate(Annotations.Symbol, r.id, sym)
+                r
+              }
+              val fvargs = fvpsyms.map { x =>
+                val r = source.Var(source.IdRef(Nil, x.name.name, source.Span.missing), source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, x)
+                Context.annotate(Annotations.Symbol, r.id, x)
+                source.ValueArg(None, r, source.Span.missing)
+              }
+              val fbpsyms = (bparams zip bps).map { (x, t) => symbols.BlockParam(Name.local(x.id.name), Some(t), x.symbol.capture, false, NoSource) }
+              val fbparams = (bparams zip fbpsyms).map { (x, sym) =>
+                val r: source.BlockParam = source.BlockParam(source.IdDef(x.id.name, source.Span.missing), Some(source.ReifiedType(sym.tpe.get)), false, source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, sym)
+                Context.annotate(Annotations.Symbol, r.id, sym)
+                r
+              }
+              val fbargs = fbpsyms.map { x =>
+                val r = source.Var(source.IdRef(Nil, x.name.name, source.Span.missing), source.Span.missing)
+                Context.annotate(Annotations.Symbol, r, x)
+                Context.annotate(Annotations.Symbol, r.id, x)
+                r
+              }
+              val ffn = fn match {
+                case source.IdTarget(id) =>
+                  val r = source.IdTarget(source.IdRef(Nil, id.name, source.Span.missing))
+                  Context.annotate(Annotations.Symbol, r.id,
+                    id.symbol match {
+                      case symbols.CallTarget(syms, impls) =>
+                        symbols.CallTarget(syms, impls) // needs to be refreshed for recursive uses
+                    })
+                  r
+                case _ => Context.panic("Implicit block argument should be an (eta-expanded) name, not an expression")
+              }
+              source.BlockLiteral(ftparams, fvparams, fbparams,
+                source.Return(source.Call(ffn, ftargs, fvargs, fbargs,
+                  source.Span.missing), source.Span.missing), source.Span.missing)
+            case _ => Context.panic("Unexpected implicit value for implicit block parameter")
+          }
+        case (a, symbols.BlockType.InterfaceType(tCons, tArgs)) =>
+          // There is nothing to do here
+          a
+      }
+    } else {
+      Context.abort("Not instantiating implicit block argument since there are errors.")
+    }
+  }
+  def instantiateImplicitValue(v: source.ValueArg, tpe: symbols.ValueType)(using Context): source.ValueArg = {
+    // There is nothing to do here
+    v
+  }
+
 }
